Moving javapattern package inside of operations, where it belongs.
