+++ astra-core/src/main/java/org/alfasoftware/astra/core/refactoring/operations/javapattern/JavaPatternReplacement.java
@@ -1,4 +1,4 @@
-package org.alfasoftware.astra.core.refactoring.javapattern;
+package org.alfasoftware.astra.core.refactoring.operations.javapattern;
 
 /**
  * This is used to annotate methods which describe the pattern to use to rewrite a matched pattern
