+++ astra-core/src/main/java/org/alfasoftware/astra/core/refactoring/operations/javapattern/Substitute.java
@@ -1,4 +1,4 @@
-package org.alfasoftware.astra.core.refactoring.javapattern;
+package org.alfasoftware.astra.core.refactoring.operations.javapattern;
 
 /**
  * This is used to annotate methods which should be subsituted as part of a {@link @JavaPattern} rewrite
