+++ astra-core/src/test/java/org/alfasoftware/astra/core/refactoring/methods/methodInvocation/InvocationChangeOutsideClassQualifiedExampleMatcher.java
@@ -1,7 +1,7 @@
 package org.alfasoftware.astra.core.refactoring.methods.methodInvocation;
 
-import org.alfasoftware.astra.core.refactoring.javapattern.JavaPattern;
-import org.alfasoftware.astra.core.refactoring.javapattern.JavaPatternReplacement;
+import org.alfasoftware.astra.core.refactoring.operations.javapattern.JavaPattern;
+import org.alfasoftware.astra.core.refactoring.operations.javapattern.JavaPatternReplacement;
 import org.alfasoftware.astra.exampleTypes.A;
